requirements.txt
easydict==1.9
hydra-core==1.1.1
joblib==1.0.1
matplotlib==3.3.4
numpy==1.22.3
omegaconf==2.1.2
open3d==0.10.0.0
polyscope==1.3.1
potpourri3d==0.0.7
protobuf==3.20.1
PyYAML==6.0
robust-laplacian==0.2.4
scikit-learn==0.23.2
scipy==1.6.2
tqdm==4.62.3
wandb==0.12.14
